insnsd.c: don't generate an inaccessible table
Some pseudo-instructions (RESB and EQU) seem to make it into the
instruction table.  This also generates an instruction table for
zero-length instructions, which of course can never actually be
accessed.  Quiet a compiler warning by simply not emitting this
useless table.  Ideally we shouldn't emit the pseudo-instructions
either, but that is a bigger change, and it's hardly a lot of memory
involved.