nsis: Inform a user about files being deleted
Since a user might have some weird access rules on NTFS
such as allow to write files but not to delete them it
might happen that uninstallation procedure fails at
deleting files but deleting registry keys still success,
and in result a user might try to uninstall nasm for second
time with registry keys corrupted.

So prompt a user with files being deleted so he can double
check what is going on.

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>