doc: include kerning tables in the metrics headers
Output the kerning tables in the metrics headers.  This will hopefully
make it possible to support kerning later (e.g. using the PostScript kshow
operator with some kind of kerning function.)

Note: the previous ordering of the metrics output was sensitive to the
particular hashing algorithm used in that version of Perl.  This
version sorts them in order to keep them stable, but which obviously
completely changes the (otherwise unmodified) widths part of the output.

Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>